api: check for update id equality when choosing previous
For whatever reason, it seems like the `createdAt` timestamp returned by
the API to the `eas` CLI when creating an update can be different from
the `createdAt` timestamp that the GraphQL API returns when listing the
same update during webhook handling. This led our webhook to choose the
same update as the "previous update" and generate useless changelogs.
Adding an ID inequality ensures we choose a more correct "previous"
target.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>